TabBar isolated from the controller.
